Comiited changes for FB
