+++ sh.exe.stackdump
@@ -1,18 +1,18 @@
 Exception: STATUS_STACK_OVERFLOW at rip=7FF99001CE47
 rax=00000000000020D0 rbx=0000000000000020 rcx=00007FF99001FC20
-rdx=0000000000351DC0 rsi=00000000FFFFAEA0 rdi=0000000000000020
+rdx=0000000000761DC0 rsi=00000000FFFFAEA0 rdi=0000000000000020
 r8 =00000000FFFFAA90 r9 =00007FF99001E8EA r10=00000000FFFF8000
 r11=00000000FFE03230 r12=00000001005E6E6C r13=00000006000564E0
 r14=00000001005E4001 r15=00000006000564C0
 rbp=00000000FFFFAEC0 rsp=00000000FFFFAA18
-program=C:\Program Files\Git\usr\bin\sh.exe, pid 7780, thread unknown (0x18B0)
+program=C:\Program Files\Git\usr\bin\sh.exe, pid 11440, thread unknown (0x2BC4)
 cs=0033 ds=002B es=002B fs=0053 gs=002B ss=002B
 Stack trace:
 Frame        Function    Args
-000FFFFAEC0  7FF99001CE47 (7FF9900190F2, 7FF99001FC20, 00000351DC0, 000FFFFAA90)
-000FFFFAEC0  7FF9900115FB (00000351DC0, 000FFFFAA90, 7FF99001E8EA, 00000000003)
+000FFFFAEC0  7FF99001CE47 (7FF9900190F2, 7FF99001FC20, 00000761DC0, 000FFFFAA90)
+000FFFFAEC0  7FF9900115FB (00000761DC0, 000FFFFAA90, 7FF99001E8EA, 00000000003)
 000FFFFAEC0  7FF9900190F2 (00040000000, 00000000003, 0018029AE40, 00000000003)
-000FFFFAEC0  001800BEF63 (000FFFFAE40, 0000000000C, 00000000188, 00000000022)
+000FFFFAEC0  001800BEF63 (000FFFFAE40, 0000000000C, 00000000198, 00000000022)
 000FFFFAEC0  001800BFDE5 (00000000000, 001004D5858, 0010040F65E, 000FFFFB208)
 00000000000  0018012A2EB (00000000000, 001004D5858, 0010040F65E, 000FFFFB208)
 00000000000  0000023E458 (001004D5858, 0010040F65E, 000FFFFB208, 00000080002)
