--- src/test/java/Utilities/Constant.java
@@ -1,20 +1,0 @@
-package Utilities;
-
-public class Constant {
-	public static final String BaseURL = "http://omdev.ca-labs.com/app/";
-	public static final String Lnk_RevQueue = "queue/reviewer";
-	public static final String Lnk_DtaAnalysys = "omReport";
-	// public static final String Username = "testuser_1";
-	// public static final String Password = "Test@123";
-	public static final String Path_TestDataLogin = "c://Users//vrca//Downloads//Maven-Failsafe-Plugin-Example-master//src//test//resources//com//testdata//";
-	public static final String File_TestDataLogin = "TestData.xlsx";
-	public static final String Path_EBITDA_PPMSimple = "C://Users//vrca//Downloads//Maven-Failsafe-Plugin-Example-master//src//test//resources//com//testdata//";
-	public static final String File_EBITDA_PPMSimple = "SIMPLE_Avg_EBITDA_PPM_by_Acquisition_Year.xlsx";
-	public static final String Path_EBITDA_PPMComptypSimple = "C://Users//vrca//Downloads//Maven-Failsafe-Plugin-Example-master//src//test//resources//com//testdata//";
-	public static final String File_EBITDA_PPMComptypSimple = "EBITDA_PPM by Company Type by Acquisition Year-31-01-2017 08-06-10.xlsx";
-	public static final String Path_EBITDA_PPMSectorSimple = "C://Users//vrca//Downloads//Maven-Failsafe-Plugin-Example-master//src//test//resources//com//testdata//";
-	public static final String File_EBITDA_PPMSectorSimple = "EBITDA PPM by GICS Sector by Acquisition Year-31-01-2017 10-41-36.xlsx";
-	public static final String Path_EBITDA_PPMExAnaSimple = "C://Users//vrca//Downloads//Maven-Failsafe-Plugin-Example-master//src//test//resources//com//testdata//";
-	public static final String File_EBITDA_PPMExAnaSimple = "SIMPLE-Median EBITDA PPM by Acquisition Year-31-01-2017 11-01-39.xlsx";
-
-}
