+++ src/test/resources/com.cucumber.testng.examples/FB_Login.feature
@@ -1,0 +1,31 @@
+#Author: your.email@your.domain.com
+#Keywords Summary :
+#Feature: List of scenarios.
+#Scenario: Business rule through list of steps with arguments.
+#Given: Some precondition step
+#When: Some key actions
+#Then: To observe outcomes or validation
+#And,But: To enumerate more Given,When,Then steps
+#Scenario Outline: List of steps for data-driven as an Examples and <placeholder>
+#Examples: Container for s table
+#Background: List of steps run before each of the scenarios
+#""" (Doc Strings)
+#| (Data Tables)
+#@ (Tags/Labels):To group Scenarios 
+#<> (placeholder)
+#""
+## (Comments)
+
+#Sample Feature Definition Template
+@tag
+Feature: Title of your feature
+	I want to use this template for my feature file
+
+@tag1
+Scenario: Login with a valid user
+    When User enters a valid user name
+    And User enters a valid password
+    And User clicks in the Login button
+    Then user shall be redirected to the FB main page
+
+
